Finish migration to JDK7
Multicatch, for-loop, switch over strings, try catch, etc.


Former-commit-id: 23dfe2996f421a8abeecd41cdb17f3887f208f46

+++ seqware-portal/src/main/java/net/sourceforge/solexatools/validation/ExperimentUpdateValidator.java
@@ -29,7 +29,7 @@ public class ExperimentUpdateValidator extends ExperimentValidator {
 
             Experiment inDbExperiment = this.getExperimentService().findByTitle(experiment.getTitle());
 
-            if (inDbExperiment != null && inDbExperiment.getSwAccession().intValue() != experiment.getSwAccession().intValue()) {
+            if (inDbExperiment != null && inDbExperiment.getSwAccession() != experiment.getSwAccession()) {
                 errors.reject("error.match.title");
             }
         }
